requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.1
Flask-SocketIO>=5.3.6
Werkzeug>=3.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
msal>=1.24.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
cryptography>=41.0.0
email-validator>=2.1.0
eventlet>=0.33.0
gunicorn>=21.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
pdfplumber>=0.10.0
python-docx>=1.0.0
websockets>=12.0
python-dateutil>=2.8.2
